Corresponds to SpecsFor.Mvc 4.0.0-rc03
Updated SpecsFor.Mvc to MVC 5.2.2.
MvcWebApp now exposes a RemoteWebDriver rather than the lower-level IWebDriver.
Improved how validation errors are detected.
Improvements to some error messages.

+++ SpecsFor.Mvc/PublishPackage.ps1
@@ -1,9 +1,17 @@
+Param(
+  [switch]$DoNotPush
+)
 #Remove existing packages
 Remove-Item *.nupkg
 #Create package
 msbuild SpecsFor.Mvc.csproj "/p:Configuration=Release"
 nuget pack SpecsFor.Mvc.csproj -Properties Configuration=Release
-#Push
-$PackageName = gci "*.nupkg"
-nuget push $PackageName
-Remove-Item *.nupkg
+if ($DoNotPush) {
+	Write-Host -ForegroundColor Yellow "Skipping 'nuget push' step!"
+}
+else {
+	#Push
+	$PackageName = gci "*.nupkg"
+	nuget push $PackageName
+	Remove-Item *.nupkg
+}
